+++ pkg/adapter/kubernetes/blueprintcr/v1/effectiveBlueprint.go
@@ -5,32 +5,14 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/cloudogu/k8s-blueprint-lib/json/entities"
-
+	crd "github.com/cloudogu/k8s-blueprint-lib/api/v1"
 	"github.com/cloudogu/k8s-blueprint-operator/v2/pkg/adapter/serializer"
 	"github.com/cloudogu/k8s-blueprint-operator/v2/pkg/domain"
 )
 
 var configKeySeparator = "/"
 
-// EffectiveBlueprint describes an abstraction of CES components that should be absent or present within one or more CES
-// instances after combining the blueprint with the blueprint mask.
-//
-// In general additions without changing the version are fine, as long as they don't change semantics. Removal or
-// renaming are breaking changes and require a new blueprint API version.
-type EffectiveBlueprint struct {
-	// Dogus contains a set of exact dogu versions which should be present or absent in the CES instance after which this
-	// blueprint was applied. Optional.
-	Dogus []entities.TargetDogu `json:"dogus,omitempty"`
-	// Components contains a set of exact component versions which should be present or absent in the CES instance after which
-	// this blueprint was applied. Optional.
-	Components []entities.TargetComponent `json:"components,omitempty"`
-	// Config is used for ecosystem configuration to be applied.
-	// Optional.
-	Config entities.TargetConfig `json:"config,omitempty"`
-}
-
-func ConvertToEffectiveBlueprintDTO(blueprint domain.EffectiveBlueprint) (EffectiveBlueprint, error) {
+func ConvertToEffectiveBlueprintDTO(blueprint domain.EffectiveBlueprint) (crd.EffectiveBlueprint, error) {
 	var errorList []error
 	convertedDogus, doguError := serializer.ConvertToDoguDTOs(blueprint.Dogus)
 	convertedComponents, componentError := serializer.ConvertToComponentDTOs(blueprint.Components)
@@ -38,17 +20,17 @@ func ConvertToEffectiveBlueprintDTO(blueprint domain.EffectiveBlueprint) (Effect
 
 	err := errors.Join(errorList...)
 	if err != nil {
-		return EffectiveBlueprint{}, fmt.Errorf("cannot convert blueprintMask to BlueprintMaskV1 DTO: %w", err)
+		return crd.EffectiveBlueprint{}, fmt.Errorf("cannot convert blueprintMask to BlueprintMaskV1 DTO: %w", err)
 	}
 
-	return EffectiveBlueprint{
+	return crd.EffectiveBlueprint{
 		Dogus:      convertedDogus,
 		Components: convertedComponents,
 		Config:     ConvertToConfigDTO(blueprint.Config),
 	}, nil
 }
 
-func ConvertToEffectiveBlueprintDomain(blueprint EffectiveBlueprint) (domain.EffectiveBlueprint, error) {
+func ConvertToEffectiveBlueprintDomain(blueprint crd.EffectiveBlueprint) (domain.EffectiveBlueprint, error) {
 	convertedDogus, doguErr := serializer.ConvertDogus(blueprint.Dogus)
 	convertedComponents, compErr := serializer.ConvertComponents(blueprint.Components)